+++ imgaug/augmenters/pooling.py
@@ -455,15 +455,10 @@ class MinPooling(_AbstractPoolingBase):
     def _pool_image(self, image, kernel_size_h, kernel_size_w):
         # TODO extend pool to support pad_mode and set it here
         #      to reflection padding
-
-        # we use cval=255 here so that the minimum is always a pixel value,
-        # even if the image was automatically padded
-        return ia.pool(
+        return ia.min_pool(
             image,
-            (kernel_size_h, kernel_size_w),
-            np.min,
-            cval=255,
-            preserve_dtype=True)
+            (max(kernel_size_h, 1), max(kernel_size_w, 1))
+        )
 
 
 class MedianPooling(_AbstractPoolingBase):